chore!: moved getRunningFeature to testKit
